Updated Text component with clientlib
