+++ ui.apps/src/main/content/jcr_root/apps/core/fd/components/form/text/v1/text/_cq_editConfig.xml
@@ -13,57 +13,55 @@
         <inplaceEditingConfig
                 jcr:primaryType="cq:Widget"
                 hideLabel="{Boolean}true"
-                name="./value"
-                textPropertyName="./value"
-                xtype="richtext">
+                textPropertyName="./value">
             <rtePlugins jcr:primaryType="nt:unstructured">
                 <footnoteplugin
                         jcr:primaryType="nt:unstructured"
                         features="*"/>
-                <links jcr:primaryType="nt:unstructured">
-                    <linkDialogConfig
-                            jcr:primaryType="nt:unstructured"
-                            height="{Long}316">
-                        <linkAttributes jcr:primaryType="cq:WidgetCollection">
-                            <linkAdvanced
-                                    jcr:primaryType="cq:Widget"
-                                    collapsed="{Boolean}true"
-                                    collapsible="{Boolean}true"
-                                    inputValue="advanced"
-                                    name="./linkdialog/cq:adhocLinkTrackingTab"
-                                    title="Link tracking"
-                                    xtype="dialogfieldset">
-                                <items jcr:primaryType="cq:WidgetCollection">
-                                    <enable
-                                            jcr:primaryType="nt:unstructured"
-                                            attribute="enabletracking"
-                                            fieldDescription="override analytics framework settings."
-                                            fieldLabel="Custom link tracking"
-                                            name="./linkdialog/cq:adhocLinkTrackingEnableTracking"
-                                            xtype="checkbox">
-                                        <listeners
-                                                jcr:primaryType="nt:unstructured"
-                                                check="function(component){var dlg=component.findParentByType('rtelinkdialog');dlg.enableSCFields(component.checked);}"/>
-                                    </enable>
-                                    <events
-                                            jcr:primaryType="nt:unstructured"
-                                            attribute="adhocevents"
-                                            fieldDescription="e.g.: event2, event7."
-                                            fieldLabel="Include SiteCatalyst events"
-                                            name="./linkdialog/cq:adhocLinkTrackingEvents"
-                                            xtype="textfield"/>
-                                    <evars
-                                            jcr:primaryType="nt:unstructured"
-                                            attribute="adhocevars"
-                                            fieldDescription="e.g.: eVar1: pagedata.url, prop4: 'const'."
-                                            fieldLabel="Include SiteCatalyst variables"
-                                            name="./linkdialog/cq:adhocLinkTrackingEvars"
-                                            xtype="textfield"/>
-                                </items>
-                            </linkAdvanced>
-                        </linkAttributes>
-                    </linkDialogConfig>
-                </links>
+<!--                <links jcr:primaryType="nt:unstructured">-->
+<!--                    <linkDialogConfig-->
+<!--                            jcr:primaryType="nt:unstructured"-->
+<!--                            height="{Long}316">-->
+<!--                        <linkAttributes jcr:primaryType="cq:WidgetCollection">-->
+<!--                            <linkAdvanced-->
+<!--                                    jcr:primaryType="cq:Widget"-->
+<!--                                    collapsed="{Boolean}true"-->
+<!--                                    collapsible="{Boolean}true"-->
+<!--                                    inputValue="advanced"-->
+<!--                                    name="./linkdialog/cq:adhocLinkTrackingTab"-->
+<!--                                    title="Link tracking"-->
+<!--                                    xtype="dialogfieldset">-->
+<!--                                <items jcr:primaryType="cq:WidgetCollection">-->
+<!--                                    <enable-->
+<!--                                            jcr:primaryType="nt:unstructured"-->
+<!--                                            attribute="enabletracking"-->
+<!--                                            fieldDescription="override analytics framework settings."-->
+<!--                                            fieldLabel="Custom link tracking"-->
+<!--                                            name="./linkdialog/cq:adhocLinkTrackingEnableTracking"-->
+<!--                                            xtype="checkbox">-->
+<!--                                        <listeners-->
+<!--                                                jcr:primaryType="nt:unstructured"-->
+<!--                                                check="function(component){var dlg=component.findParentByType('rtelinkdialog');dlg.enableSCFields(component.checked);}"/>-->
+<!--                                    </enable>-->
+<!--                                    <events-->
+<!--                                            jcr:primaryType="nt:unstructured"-->
+<!--                                            attribute="adhocevents"-->
+<!--                                            fieldDescription="e.g.: event2, event7."-->
+<!--                                            fieldLabel="Include SiteCatalyst events"-->
+<!--                                            name="./linkdialog/cq:adhocLinkTrackingEvents"-->
+<!--                                            xtype="textfield"/>-->
+<!--                                    <evars-->
+<!--                                            jcr:primaryType="nt:unstructured"-->
+<!--                                            attribute="adhocevars"-->
+<!--                                            fieldDescription="e.g.: eVar1: pagedata.url, prop4: 'const'."-->
+<!--                                            fieldLabel="Include SiteCatalyst variables"-->
+<!--                                            name="./linkdialog/cq:adhocLinkTrackingEvars"-->
+<!--                                            xtype="textfield"/>-->
+<!--                                </items>-->
+<!--                            </linkAdvanced>-->
+<!--                        </linkAttributes>-->
+<!--                    </linkDialogConfig>-->
+<!--                </links>-->
                 <misctools
                         jcr:primaryType="nt:unstructured"
                         features="*"/>
