+++ ui.apps/src/main/content/jcr_root/apps/core/fd/components/form/text/v1/text/_cq_template.xml
@@ -2,4 +2,4 @@
 <jcr:root xmlns:jcr="http://www.jcp.org/jcr/1.0" xmlns:nt="http://www.jcp.org/jcr/nt/1.0" xmlns:cq="http://www.day.com/jcr/cq/1.0"
     jcr:primaryType="nt:unstructured"
     jcr:title="Text"
-    fieldType="text"/>
+    fieldType="plain-text"/>
