--- ui.apps/src/main/content/jcr_root/apps/core/fd/components/form/text/v1/text/clientlibs/editor/.content.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<jcr:root xmlns:cq="http://www.day.com/jcr/cq/1.0" xmlns:jcr="http://www.jcp.org/jcr/1.0"
-    jcr:primaryType="cq:ClientLibraryFolder"
-    categories="[core.forms.components.text.v1.editor]"
-    dependencies="[core.forms.components.base.v1.editor, core.forms.components.commons.v1.editor.utils]"/>
