+++ ui.apps/src/main/content/jcr_root/apps/core/fd/components/form/text/v1/text/clientlibs/site/.content.xml
@@ -2,4 +2,4 @@
 <jcr:root xmlns:cq="http://www.day.com/jcr/cq/1.0" xmlns:jcr="http://www.jcp.org/jcr/1.0"
           jcr:primaryType="cq:ClientLibraryFolder"
           categories="[core.forms.components.text.v1.runtime]"
-          dependencies="[jquery,core.forms.components.container.v1.runtime]"/>
+          dependencies="[core.forms.components.container.v1.runtime]"/>
