+++ ui.apps/src/main/content/jcr_root/apps/core/fd/components/form/text/v1/text/clientlibs/site/css/textview.css
@@ -13,12 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  ******************************************************************************/
-(function($) {
-    "use strict";
 
-    var EDIT_DIALOG = ".cmp-adaptiveform-text__editdialog",
-        Utils = window.CQ.FormsCoreComponents.Utils.v1;
+.cmp-adaptiveform-text {
 
-    Utils.initializeEditDialog(EDIT_DIALOG)();
+}
 
-})(jQuery);
+.cmp-adaptiveform-text__widget {
+
+}
+
+[data-cmp-visible=false] {
+    display: none;
+}
