+++ ui.apps/src/main/content/jcr_root/apps/core/fd/components/form/text/v1/text/clientlibs/site/js.txt
@@ -15,4 +15,4 @@
 ###############################################################################
 
 #base=js
-text.js
+textview.js
