+++ ui.apps/src/main/content/jcr_root/apps/core/fd/components/form/text/v1/text/text.html
@@ -14,14 +14,17 @@
   ~ limitations under the License.
   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/-->
 <div data-sly-use.textModel="com.adobe.cq.forms.core.components.models.form.Text"
+     data-sly-use.formstructparser="com.adobe.cq.forms.core.components.models.form.FormStructureParser"
      data-sly-use.component="com.adobe.cq.wcm.core.components.models.Component"
      data-sly-use.templates="core/wcm/components/commons/v1/templates.html"
      data-sly-test.text="${textModel.value}"
-     data-cmp-data-layer="${textModel.data.json}"
      data-cmp-is="adaptiveFormText"
      id="${component.id}"
-     class="cmp-text">
-    <p class="cmp-text__paragraph"
-       data-sly-unwrap="${textModel.isRichText}">${text @ context = textModel.isRichText ? 'html' : 'text'}</p>
+     data-cmp-adaptiveformcontainer-path="${formstructparser.formContainerPath}"
+     data-cmp-visible="${textModel.visible}"
+     class="cmp-adaptiveform-text">
+    <div class="cmp-adaptiveform-text__widget">
+        <div data-sly-unwrap="${textModel.isRichText}">${text @ context = textModel.isRichText ? 'html' : 'text'}</div>
+    </div>
 </div>
-<sly data-sly-call="${templates.placeholder @ isEmpty = !text, classAppend='cmp-text'}"></sly>
+<sly data-sly-call="${templates.placeholder @ isEmpty = !text, classAppend='cmp-adaptiveform-text'}"></sly>
