+++ ui.apps/src/main/content/jcr_root/apps/core/fd/components/form/textinput/v1/textinput/clientlibs/site/css/textinputview.css
@@ -37,3 +37,5 @@
 .cmp-adaptiveform-textinput__questionmark {
 
 }
+
+[data-cmp-visible=false]
